Remove lib_player_list, move funcs to lib_player
