+++ deploy/lib/control/HomepageController.php
@@ -1,8 +1,6 @@
 <?php
 namespace NinjaWars\core\control;
 
-require_once(LIB_ROOT.'control/lib_player_list.php'); // Used for member_counts
-
 use NinjaWars\core\data\Message;
 use \Player as Player;
