+++ deploy/www/api.php
@@ -1,5 +1,4 @@
 <?php
-require_once(LIB_ROOT.'control/lib_player_list.php');
 require_once(LIB_ROOT.'control/lib_api.php');
 // How to call:  http://nw.local/api.php?type=char_search&jsoncallback=alert&term=tchalvak&limit=10
 // http://nw.local/api.php?type=facebook_login_sync&jsoncallback=alert
